requirements.txt
Django==3.1.7
django-environ==0.4.5
django-tinymce==3.2.0
factory-boy==3.2.0
Pillow==8.1.0
psycopg2-binary==2.8.6
